+++ server/db/049-updateclienttables.sql
@@ -3,7 +3,7 @@ DO $migration$
     DECLARE lastSuccessfulMigrationId INT;
 BEGIN
     -- The migration ID of this file
-    migrationId := 48;
+    migrationId := 49;
 
     -- Get the migration ID of the last file to be successfully run
     SELECT MAX(id) INTO lastSuccessfulMigrationId
@@ -11,7 +11,8 @@ BEGIN
 
     -- Only execute this script if its migration ID is next after the last successful migration ID
     IF migrationId - lastSuccessfulMigrationId = 1 THEN
-        ALTER TABLE clients ALTER COLUMN incident_categories SET DEFAULT '{"Overdose Event","Safer Use","Conflict Resolution","Test"}';
+        ALTER TABLE clients_extension ADD COLUMN city text ;
+        ALTER TABLE clients_extension ADD COLUMN project text ;
 
         -- Update the migration ID of the last file to be successfully run to the migration ID of this file
         INSERT INTO migrations (id)