+++ CHANGELOG.md
@@ -11,9 +11,16 @@ the code was deployed.
 
 ## Unreleased
 
+### Changed 
+
+- Updated the .env.example and README for Buttons (CU-86duebt59)
+
+## [13.12.0] - 2024-09-12
+
 ### Added
 
-- Added migration script to change the default value in the `incident_categories` column to `{"Overdose Event","No One Inside","Occupant Responded","None of the Above"}` (CU-86du7hv9b)
+- Added city, projects to clients_extension (CU-86du6jp33)
+- Added migration script to change the default value in the `incident_categories` column to `{"Overdose Event","Safer Use","Conflict Resolution","Test"}` (CU-86du7hv9b)
 - Added migration script to insert `organization`, `funder`, and `postal_code` columns into the `clients_extention` table (CU-86du6jp33).
 - New REST API routes (still require test cases before further use) (CU-86dugzyzh).
 - Added `ClientExtension.js` with class `ClientExtension` to be used in adding and editing clients (CU-86du23zth).
@@ -811,9 +818,10 @@ the code was deployed.
 
 - Initial pi.
 
-[unreleased]: https://github.com/bravetechnologycoop/BraveButtons/compare/v13.11.0...HEAD
+[unreleased]: https://github.com/bravetechnologycoop/BraveButtons/compare/v13.12.0...HEAD
+[13.12.0]: https://github.com/bravetechnologycoop/BraveButtons/compare/v13.11.0...v13.12.0
 [13.11.0]: https://github.com/bravetechnologycoop/BraveButtons/compare/v13.10.0...v13.11.0
-[13.10.0]: https://github.com/bravetechnologycoop/BraveButtons/compare/v13.9.0...v13.10.0
+[13.10.0]: https://github.com/bravetechnologycoop/BraveButtons/compare/v13.9.0...v13.10.0git
 [13.9.0]: https://github.com/bravetechnologycoop/BraveButtons/compare/v13.8.0...v13.9.0
 [13.8.0]: https://github.com/bravetechnologycoop/BraveButtons/compare/v13.7.1...v13.8.0
 [13.7.1]: https://github.com/bravetechnologycoop/BraveButtons/compare/v13.7.0...v13.7.1